requirements.txt
pyqt5~=5.15.9
requests~=2.32.4
pyqt-fluent-widgets[full]~=1.9.0
pyyaml
